# internal/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Skifity API
  version: development
  summary: Self-hosted apps, powered by Kubernetes.
  description: |
    The API the panel, the `skifity` CLI and the MCP server all use. Anything the
    panel can do, a request here can do.

    **Authentication.** Send an API token as `Authorization: Bearer skf_...`.
    Create one in the panel under Account, then API tokens. A browser session
    works too; it is the `skifity_session` cookie (`__Host-skifity_session` on
    HTTPS), and a request that changes something then also carries the
    session's CSRF token in the `X-Skifity-CSRF` header. A token needs no CSRF
    header. A few actions refuse a token and want a person who has just typed
    their password; those list only the session scheme.

    **Answers.** Lists come back as `{"items": [...], "total": n}`. Every
    failure has the same shape: an `error` object with a stable `code`, a
    `title`, and the `cause`, `impact` and `fix` in words to act on.

    **Streams.** The event stream, and an app's logs followed live, are
    server-sent events (`text/event-stream`).

    **Roles.** A team member is an owner, an admin, a member or a viewer. A
    viewer reads everything a member can and changes nothing. Routes marked as
    for panel administrators want an account that administers the whole
    panel, whatever its role in a team.
  license:
    name: Apache License 2.0
    identifier: Apache-2.0
servers:
  - url: /
...

...
security:
  - bearerToken: []
  - sessionCookie: []
...
paths:
...
  /api/teams/{teamID}:
    parameters:
      - $ref: '#/components/parameters/teamID'
    get:
      operationId: getTeam
      tags: [Teams]
      summary: A team, and your place in it
      responses:
        '200':
          description: The team, with your role and any limit to projects.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Team'}
        default: {$ref: '#/components/responses/Error'}
    patch:
      operationId: updateTeam
      tags: [Teams]
      summary: Rename a team, or require a second factor in it
      description: >-
        Admin. The slug does not change, because it is part of every namespace
        name. Requiring a second factor is refused to somebody who signed in
        without one, who would be locked out by it.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UpdateTeamRequest'}
      responses:
        '200':
          description: The team as it is now.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Team'}
        default: {$ref: '#/components/responses/Error'}
...
  /api/projects/{projectID}:
    parameters:
      - $ref: '#/components/parameters/projectID'
    get:
      operationId: getProject
      tags: [Projects]
      summary: A project
      responses:
        '200':
          description: The project.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Project'}
        default: {$ref: '#/components/responses/Error'}
    patch:
      operationId: updateProject
      tags: [Projects]
      summary: Rename a project, or change its description
      description: Member.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UpdateProjectRequest'}
      responses:
        '200':
          description: The project as it is now.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Project'}
        default: {$ref: '#/components/responses/Error'}
    delete:
      operationId: deleteProject
      tags: [Projects]
      summary: Delete a project and everything in it
      description: Admin. Every environment's namespace is removed from the cluster with it.
      responses:
        '200': {$ref: '#/components/responses/OK'}
        default: {$ref: '#/components/responses/Error'}
  /api/projects/{projectID}/environments:
    parameters:
      - $ref: '#/components/parameters/projectID'
    get:
      operationId: listEnvironments
      tags: [Projects]
      summary: A project's environments
      responses:
        '200':
          description: The environments, previews included.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/EnvironmentList'}
        default: {$ref: '#/components/responses/Error'}
    post:
      operationId: createEnvironment
      tags: [Projects]
      summary: Create an environment
      description: Member. It is a namespace of its own, with its own quota and network policy.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateEnvironmentRequest'}
      responses:
        '201':
          description: The environment.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Environment'}
        default: {$ref: '#/components/responses/Error'}
...
  /api/environments/{envID}:
    parameters:
      - $ref: '#/components/parameters/envID'
    get:
      operationId: getEnvironment
      tags: [Projects]
      summary: An environment
      responses:
        '200':
          description: The environment.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Environment'}
        default: {$ref: '#/components/responses/Error'}
    patch:
      operationId: updateEnvironment
      tags: [Projects]
      summary: Change how strictly an environment confines its apps, or what its previews copy
      description: Admin.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UpdateEnvironmentRequest'}
      responses:
        '200':
          description: The environment as it is now.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Environment'}
        default: {$ref: '#/components/responses/Error'}
    delete:
      operationId: deleteEnvironment
      tags: [Projects]
      summary: Delete an environment and everything in it
      description: Admin. Its namespace is removed from the cluster.
      responses:
        '200': {$ref: '#/components/responses/OK'}
        default: {$ref: '#/components/responses/Error'}
...
  /api/environments/{envID}/apps:
    parameters:
      - $ref: '#/components/parameters/envID'
    get:
      operationId: listApps
      tags: [Apps]
      summary: The apps in an environment
      responses:
        '200':
          description: The apps.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AppList'}
        default: {$ref: '#/components/responses/Error'}
    post:
      operationId: createApp
      tags: [Apps]
      summary: Create an app
      description: >-
        Member. From a Git repository, an image, or code sent afterwards with
        `skifity up`. Variables, databases and processes given here exist
        before the first deploy, so the app never starts once without them.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateAppRequest'}
      responses:
        '201':
          description: The app, and whatever was made with it.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CreatedApp'}
        default: {$ref: '#/components/responses/Error'}
...
  /api/environments/{envID}/databases:
    parameters:
      - $ref: '#/components/parameters/envID'
    get:
      operationId: listDatabases
      tags: [Databases]
      summary: The databases in an environment
      responses:
        '200':
          description: The databases.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DatabaseList'}
        default: {$ref: '#/components/responses/Error'}
    post:
      operationId: createDatabase
      tags: [Databases]
      summary: Create a managed database
      description: Member. Its operator is installed on first use. It answers at once, while the database is still being made.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateDatabaseRequest'}
      responses:
        '202':
          description: The database, being created.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Database'}
        default: {$ref: '#/components/responses/Error'}

  # --- apps ---
...
  /api/apps/{appID}/jobs:
    parameters:
      - $ref: '#/components/parameters/appID'
    get:
      operationId: listAppJobs
      tags: [Apps]
      summary: An app's scheduled commands
      responses:
        '200':
          description: The scheduled commands.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AppJobList'}
        default: {$ref: '#/components/responses/Error'}
    post:
      operationId: createAppJob
      tags: [Apps]
      summary: Run a command on a schedule
      description: Admin. Five cron fields, in UTC.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AppJobRequest'}
      responses:
        '201':
          description: The scheduled command.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AppJob'}
        default: {$ref: '#/components/responses/Error'}
...
  /api/apps/{appID}/variables:
    parameters:
      - $ref: '#/components/parameters/appID'
    get:
      operationId: listVariables
      tags: [Variables]
      summary: An app's variables
      description: A secret's value, and its preview value, are never returned.
      responses:
        '200':
          description: The variables.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/VariableList'}
        default: {$ref: '#/components/responses/Error'}
    put:
      operationId: setVariable
      tags: [Variables]
      summary: Set one variable
      description: >-
        Member. A runtime variable is rolled out; a build-time one takes effect
        at the next build, and the answer says which.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SetVariableRequest'}
      responses:
        '200':
          description: The variable, without its value.
          content:
            application/json:
              schema:
                type: object
                required: [variable, requires_rebuild]
                properties:
                  variable: {$ref: '#/components/schemas/Variable'}
                  requires_rebuild: {type: boolean}
        default: {$ref: '#/components/responses/Error'}
...
  /api/apps/{appID}/domains:
    parameters:
      - $ref: '#/components/parameters/appID'
    get:
      operationId: listDomains
      tags: [Domains]
      summary: The hostnames an app answers on
      responses:
        '200':
          description: The domains, each of your own with where its DNS has to point.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DomainList'}
        default: {$ref: '#/components/responses/Error'}
    post:
      operationId: addDomain
      tags: [Domains]
      summary: Add a domain
      description: >-
        Member. A certificate is requested once its DNS points here. The panel's own hostname is refused.
        A domain whose DNS does not point here yet is still added: the answer says what its DNS said.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AddDomainRequest'}
      responses:
        '201':
          description: The domain, pending until it is served, and what its DNS said when it was added.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AddedDomain'}
        default: {$ref: '#/components/responses/Error'}
...
  /api/apps/{appID}/processes:
    parameters:
      - $ref: '#/components/parameters/appID'
    get:
      operationId: listProcesses
      tags: [Scaling]
      summary: An app's other processes
      description: The workers and clocks that run beside it on the same image.
      responses:
        '200':
          description: The processes, with how many instances are ready when the cluster says.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ProcessList'}
        default: {$ref: '#/components/responses/Error'}
  /api/apps/{appID}/processes/{process}:
    parameters:
      - $ref: '#/components/parameters/appID'
      - $ref: '#/components/parameters/process'
    put:
      operationId: setProcess
      tags: [Scaling]
      summary: Add a process, or change its command or instances
      description: Member. Up to ten per app.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SetProcessRequest'}
      responses:
        '200':
          description: The process as it is now.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AppProcess'}
        default: {$ref: '#/components/responses/Error'}
    delete:
      operationId: deleteProcess
      tags: [Scaling]
      summary: Stop and remove a process
      description: Member.
      responses:
        '204':
          description: Removed.
        default: {$ref: '#/components/responses/Error'}
...
  /api/apps/{appID}/ports:
    parameters:
      - $ref: '#/components/parameters/appID'
    get:
      operationId: listPorts
      tags: [Domains]
      summary: The ports an app takes connections on that are not HTTP
      description: Viewer. Each is open on every server, at its public port.
      responses:
        '200':
          description: The ports, with the team's servers' addresses at each.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AppPortList'}
        default: {$ref: '#/components/responses/Error'}
    post:
      operationId: addPort
      tags: [Domains]
      summary: Open a TCP or UDP port on every server
      description: >-
        Member. A public port and protocol are open for one app on the whole
        panel; one that is taken is refused. The firewall does not apply to
        these, because it reads HTTP.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AddPortRequest'}
      responses:
        '201':
          description: The port, and where it is reached.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AppPortWithAddresses'}
        default: {$ref: '#/components/responses/Error'}
...
  /api/apps/{appID}/files:
    parameters:
      - $ref: '#/components/parameters/appID'
    get:
      operationId: listFiles
      tags: [Variables]
      summary: The files an app reads
      description: Viewer. Each is mounted read-only at its path in every container the app runs. A secret file's content is never sent back.
      responses:
        '200':
          description: The files, by path.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AppFileList'}
        default: {$ref: '#/components/responses/Error'}
    put:
      operationId: setFile
      tags: [Variables]
      summary: Save a file at a path, or replace the one there
      description: >-
        Member. Rolls the app out, never rebuilds it. At most 256 KiB, and an
        app's files together at most 900 KiB. Leaving is_secret out keeps a
        secret file secret.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SetFileRequest'}
      responses:
        '200':
          description: The file as saved, without its content.
          content:
            application/json:
              schema:
                type: object
                required: [file]
                properties:
                  file: {$ref: '#/components/schemas/AppFile'}
        default: {$ref: '#/components/responses/Error'}
...
  /api/apps/{appID}/volumes:
    parameters:
      - $ref: '#/components/parameters/appID'
    get:
      operationId: listVolumes
      tags: [Volumes]
      summary: An app's volumes
      responses:
        '200':
          description: The volumes.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/VolumeList'}
        default: {$ref: '#/components/responses/Error'}
    post:
      operationId: createVolume
      tags: [Volumes]
      summary: Attach a volume
      description: Member.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateVolumeRequest'}
      responses:
        '201':
          description: The volume.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Volume'}
        default: {$ref: '#/components/responses/Error'}
...
components:
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
      bearerFormat: skf_...
      description: >-
        An API token, made in the panel under Account, then API tokens. It is
        made for one team, can do no more than its owner's role in that team,
        and may be narrowed further with scopes. It needs no CSRF header.
    sessionCookie:
      type: apiKey
      in: cookie
      name: skifity_session
      description: >-
        A browser session, set by signing in. On HTTPS the cookie is
        `__Host-skifity_session`. A request that changes something also sends
        the session's CSRF token, which signing in returns as csrf_token, in
        the X-Skifity-CSRF header.
  parameters:
    teamID: {name: teamID, in: path, required: true, description: 'A team''s id.', schema: {type: string}}
...
    projectID: {name: projectID, in: path, required: true, description: 'A project''s id.', schema: {type: string}}
    envID: {name: envID, in: path, required: true, description: 'An environment''s id.', schema: {type: string}}
...
    appID: {name: appID, in: path, required: true, description: 'An app''s id.', schema: {type: string}}
...
  responses:
    Error:
      description: >-
        What went wrong, what it means and what to do. The code is stable; the
        sentences are English, and args carries the values in them so the
        panel can say the same thing in another language.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
    OK:
      description: Done. There is nothing else to say.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/OK'}
...
  schemas:
...
    Team:
      type: object
      required: [id, name, slug, created_at, require_strong_auth]
      properties:
        id: {type: string}
        name: {type: string}
        slug: {type: string}
        created_at: {type: string, format: date-time}
        role: {$ref: '#/components/schemas/Role'}
        scoped: {type: boolean, description: 'Your membership is limited to some of the team''s projects.'}
        projects:
          type: array
          description: The projects it is limited to.
          items: {type: string}
        require_strong_auth: {type: boolean, description: 'Everybody signs in with a second factor or single sign-on.'}
...
    UpdateTeamRequest:
      type: object
      additionalProperties: false
      properties:
        name: {type: string, description: 'Unchanged when empty.'}
        require_strong_auth: {type: boolean}
...
    Project:
      type: object
      required: [id, team_id, name, slug, description, created_at, updated_at]
      properties:
        id: {type: string}
        team_id: {type: string}
        name: {type: string}
        slug: {type: string}
        description: {type: string}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
...
    UpdateProjectRequest:
      type: object
      additionalProperties: false
      properties:
        name: {type: string}
        description: {type: string}
    Environment:
      type: object
      required: [id, project_id, name, slug, kind, namespace, pod_security, preview_stack, created_at]
      properties:
        id: {type: string}
        project_id: {type: string}
        name: {type: string}
        slug: {type: string}
        kind: {type: string, enum: [standard, preview]}
        namespace: {type: string, description: 'The Kubernetes namespace. Shown under Advanced in the panel.'}
        source_ref: {type: string, description: 'The pull request a preview is of.'}
        pod_security: {type: string, enum: [restricted, baseline], description: 'The Pod Security level its namespace enforces.'}
        from_fork: {type: boolean, description: 'A preview of a pull request from a fork, which is given no secret.'}
        preview_stack: {type: boolean, description: 'A preview of one app here copies every app here.'}
        created_at: {type: string, format: date-time}
    EnvironmentList:
      type: object
      required: [items, total]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/Environment'}}
        total: {type: integer}
...
    CreateEnvironmentRequest:
      type: object
      required: [name]
      additionalProperties: false
      properties:
        name: {type: string}
    UpdateEnvironmentRequest:
      type: object
      additionalProperties: false
      properties:
        pod_security: {type: string, enum: [restricted, baseline]}
        preview_stack: {type: boolean}
...
    AppList:
      type: object
      required: [items, total]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/App'}}
        total: {type: integer}
...
    CreateAppRequest:
      type: object
      required: [name]
      additionalProperties: false
      properties:
        name: {type: string}
        source_type:
          type: string
          enum: [git, image, upload]
          description: image when an image is given, git otherwise.
        git_source_id: {type: string}
        repo_url: {type: string, description: 'Needed for git.'}
        branch: {type: string}
        root_dir: {type: string}
        builder: {type: string, default: auto}
        dockerfile_path: {type: string}
        build_command: {type: string}
        static_dir: {type: string}
        image: {type: string, description: 'Needed for image, such as nginx:1.27.'}
        port: {type: integer}
        health_path: {type: string}
        health_check:
          type: string
          enum: [http, tcp, none]
          description: http when a health_path is given, tcp otherwise. http needs a health_path.
        health_start_seconds: {type: integer, minimum: 10, maximum: 1800, default: 120}
        health_timeout_seconds: {type: integer, minimum: 1, maximum: 60, default: 3}
        start_command: {type: string}
        release_command: {type: string}
        preview_seed: {type: string}
        watch_paths: {type: string}
        deploy_trigger: {type: string, enum: [branch, tag]}
        tag_pattern: {type: string}
        internal: {type: boolean}
        deploy: {type: boolean, description: 'Deploy it straight away.'}
        variables:
          type: object
          description: Set before the first deploy. Each is decided secret or not by the same rule as everywhere else.
          additionalProperties: {type: string}
        databases:
          type: array
          maxItems: 4
          description: Created in the same environment and linked before the first deploy.
          items:
            type: object
            required: [engine]
            additionalProperties: false
            properties:
              engine: {type: string, enum: [postgres, mysql, mariadb, mongodb, redis, valkey, dragonfly, clickhouse, memcached]}
              variable: {type: string, description: 'The variable the app reads the connection from. The engine''s default when left out.'}
        processes:
          type: array
          description: Workers and the like, to run beside it on its image.
          items:
            type: object
            required: [name, command]
            additionalProperties: false
            properties:
              name: {type: string}
              command: {type: string}
              instances: {type: integer, default: 1}
    CreatedApp:
      type: object
      required: [app, webhook]
      properties:
        app: {$ref: '#/components/schemas/App'}
        webhook: {$ref: '#/components/schemas/WebhookStatus'}
        databases:
          type: array
          description: What happened to each database asked for. One that failed does not undo the app.
          items:
            type: object
            required: [engine, variable]
            properties:
              engine: {type: string}
              variable: {type: string}
              database_id: {type: string}
              name: {type: string}
              error: {type: string}
        deployment: {$ref: '#/components/schemas/Deployment'}
...